adding regression changes in rebase PR
